feat: deleting and starting again
