+++ tests/test_web2citwrapper_comm.py
@@ -3,5 +3,10 @@ from web2citwrapper import comm
 
 
 def test_no_url():
-    with raises(comm.URLMissingError):
+    with raises(comm.ParametersMissinError):
         comm.get()
+
+
+def test_error_url():
+    with raises(comm.Web2CitError):
+        comm.get({'ulr': 'invalid'})
