--- tests/vcr_cassetes/domain-test-invalid.yml
@@ -1,44 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate, br
-      Connection:
-      - keep-alive
-      User-Agent:
-      - python-requests/2.28.0
-    method: GET
-    uri: https://web2cit.toolforge.org/translate?format=json&domain=notexists.com.asdfg&tests=true
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAAAw3GwRGAIAwEwFauAMuxgRMDZgYSJgTr133tyWiSmMwH29aUolXlPkC7YW4CvtTO
-        qwtq+EDKmJ0p8Pi/EsWtatvBVLf1AaI+Ds9QAAAA
-    headers:
-      Access-Control-Allow-Origin:
-      - '*'
-      Connection:
-      - keep-alive
-      Content-Encoding:
-      - gzip
-      Content-Language:
-      - en
-      Content-Type:
-      - text/html; charset=utf-8
-      Date:
-      - Tue, 26 Jul 2022 02:10:26 GMT
-      ETag:
-      - W/"50-XsGPUn3gAIXLk3i++v99ksrEDn0"
-      Server:
-      - nginx/1.14.2
-      Transfer-Encoding:
-      - chunked
-      X-Powered-By:
-      - Express
-    status:
-      code: 404
-      message: Not Found
-version: 1
