--- tests/vcr_cassetes/domain-test-valid.yml
@@ -1,63 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate, br
-      Connection:
-      - keep-alive
-      User-Agent:
-      - python-requests/2.28.0
-    method: GET
-    uri: https://web2cit.toolforge.org/translate?format=json&domain=www.lavoz.com.ar&tests=true
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAAA72Tv07DMBDGX8XyiOIGGChk61iJgQGxVB2uyaWx5Pgs22lEQx6pUx+hL8alLW0Y
-        kBADUqIovj/f/T7bi046iJXMZOrI6KhzSGGjyaqc37JBHzGo3Dc5fwwochR0PsSdLjQpbSN6hwa8
-        qrXVIXoKyqEnhUZRyalgdKhJGVIe8wq2lMpEegyNiUFmi05GrJ2BiDLr+kSWGk1xClioeVFqTnh9
-        d8hl1ETXRK6SLa4crFEuE64PxyWLbXDA2jPPGOYYCzl57nHbJ508t4s6cmzca7bRh70VG7S4JQOW
-        qTzYwy4TBsQZeEgYiA97cUUWIC7QiRioBRoxohaGxIma6z7EC5nDbrB4PPbf1X+vPeh9WTF5nD49
-        3F+f6cgb16x4tyos5vabQ88g3mgrCqabDxuuyY8RfgifFe9GAuzuuhn2bew/hnGz09+ldtnz7JXH
-        kk9CFaMLWZq2bTsxsKHtJKd6Av5fjy5fF3bA8jg3N7JffgIc88Y+QQMAAA==
-    headers:
-      Access-Control-Allow-Origin:
-      - '*'
-      Connection:
-      - keep-alive
-      Content-Encoding:
-      - gzip
-      Content-Language:
-      - en
-      Content-Security-Policy-Report-Only:
-      - 'default-src ''self'' ''unsafe-eval'' ''unsafe-inline'' blob: data: filesystem:
-        mediastream: *.toolforge.org wikibooks.org *.wikibooks.org wikidata.org *.wikidata.org
-        wikimedia.org *.wikimedia.org wikinews.org *.wikinews.org wikipedia.org *.wikipedia.org
-        wikiquote.org *.wikiquote.org wikisource.org *.wikisource.org wikiversity.org
-        *.wikiversity.org wikivoyage.org *.wikivoyage.org wiktionary.org *.wiktionary.org
-        *.wmcloud.org *.wmflabs.org wikimediafoundation.org mediawiki.org *.mediawiki.org
-        wss://web2cit.toolforge.org; report-uri https://csp-report.toolforge.org/collect;'
-      Content-Type:
-      - application/json; charset=utf-8
-      Date:
-      - Tue, 26 Jul 2022 02:09:32 GMT
-      ETag:
-      - W/"341-wBT1N3Pp6zOlarnbgMhPrwJBdF8"
-      Permissions-Policy:
-      - interest-cohort=()
-      Server:
-      - nginx/1.14.2
-      Strict-Transport-Security:
-      - max-age=31622400
-      Transfer-Encoding:
-      - chunked
-      X-Clacks-Overhead:
-      - GNU Terry Pratchett
-      X-Powered-By:
-      - Express
-    status:
-      code: 200
-      message: OK
-version: 1
