--- tests/vcr_cassetes/page-check-test.yml
@@ -1,60 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate, br
-      Connection:
-      - keep-alive
-      User-Agent:
-      - python-requests/2.28.0
-    method: GET
-    uri: https://web2cit.toolforge.org/translate?url=https://mediawiki.org&format=json&tests=true
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAAA1WPMQrDMAxF76IxmGT3DTp0K3QIGRyixKKOY2wZU0zuXoV2qNEmvSfpjxWCYQsa
-        BlAQMWXHCfRYgXEPzjCCrqeCldAt34E3uzSBBHi8A4p2ZA6ZxYKCczAbwnSqCj+QiV1L3XEh86QX
-        NVzIs6Nkcbn5dmcp/X4ZRYz+iFtjOeO3fJ38fwO9MFIKbMRVnrXMIelhaPZciSU8Y/SCdB2c0wcZ
-        5LRvDwEAAA==
-    headers:
-      Access-Control-Allow-Origin:
-      - '*'
-      Connection:
-      - keep-alive
-      Content-Encoding:
-      - gzip
-      Content-Language:
-      - en
-      Content-Security-Policy-Report-Only:
-      - 'default-src ''self'' ''unsafe-eval'' ''unsafe-inline'' blob: data: filesystem:
-        mediastream: *.toolforge.org wikibooks.org *.wikibooks.org wikidata.org *.wikidata.org
-        wikimedia.org *.wikimedia.org wikinews.org *.wikinews.org wikipedia.org *.wikipedia.org
-        wikiquote.org *.wikiquote.org wikisource.org *.wikisource.org wikiversity.org
-        *.wikiversity.org wikivoyage.org *.wikivoyage.org wiktionary.org *.wiktionary.org
-        *.wmcloud.org *.wmflabs.org wikimediafoundation.org mediawiki.org *.mediawiki.org
-        wss://web2cit.toolforge.org; report-uri https://csp-report.toolforge.org/collect;'
-      Content-Type:
-      - application/json; charset=utf-8
-      Date:
-      - Mon, 25 Jul 2022 14:55:23 GMT
-      ETag:
-      - W/"10f-vLj9nbDmCdpkjGHephree3CDZMc"
-      Permissions-Policy:
-      - interest-cohort=()
-      Server:
-      - nginx/1.14.2
-      Strict-Transport-Security:
-      - max-age=31622400
-      Transfer-Encoding:
-      - chunked
-      X-Clacks-Overhead:
-      - GNU Terry Pratchett
-      X-Powered-By:
-      - Express
-    status:
-      code: 200
-      message: OK
-version: 1
