+++ web2citwrapper/__init__.py
@@ -1,3 +1,2 @@
-from .page import Page
 from .comm import *
-from .domain import Domain
+from .url import URL
