+++ web2citwrapper/comm.py
@@ -5,14 +5,24 @@ URL = 'https://web2cit.toolforge.org/translate?'
 PARAMETERS = {'format': 'json'}
 
 
-class URLMissingError(Exception):
+class ParametersMissinError(Exception):
+    """Raised when parameters is not provided"""
     pass
 
 
-def get(parameters=None):
+class Web2CitError(Exception):
+    """Raised when the web2cit service returns an error"""
+    pass
+
+
+def get(parameters: dict = None) -> dict:
     """Retrieve a simple endpoint to web2cit"""
     if parameters is None:
-        raise URLMissingError
+        raise ParametersMissinError
 
     request = requests.get(URL, {**PARAMETERS, **parameters})
-    return request.json()
+    json = request.json()
+    if 'error' in json.keys():
+        raise Web2CitError(json['error']['message'])
+
+    return json
